Kerberos key server code
Simple Kerberos key server implementation consisting of a KeyServer that
takes in requests from UserClient and UserServer. UserClient makes a
request to the KeyServer for a session. Once session request is granted,
a session key is sent to the UserClient to be used for communication
with the UserServer. Client and server talk to each other by encrypting
and decrypting messages using this session key.